some refactoring & small reorganizations
