+++ src/components/families/family-sidebar/FamilySidebar.tsx
@@ -32,9 +32,9 @@ import { UsersContext } from "context/UsersContext";
 
 import { FamilyFormData } from "../family-form/types";
 import EnrolmentForm from "./enrolment-form";
-import FamilySidebarEnrolmentCard from "./family-sidebar-enrolment-card";
 import FamilySidebarForm, { familySidebarFormId } from "./family-sidebar-form";
 import InteractionCard from "./interaction-card";
+import PreviousEnrolmentCard from "./previous-enrolment-card";
 
 const DRAWER_WIDTH = 416;
 
@@ -364,24 +364,20 @@ const FamilySidebar = ({
           <Divider />
         </Box>
 
-        <Box position="relative">
-          <Box>
-            <Typography variant="h3" className={classes.heading}>
-              Previous Enrolments
-            </Typography>
-            {family.enrolments
-              .filter((enrolment) => !enrolment.session.active)
-              .map((enrolment) => (
-                <FamilySidebarEnrolmentCard
-                  key={enrolment.id}
-                  enrolment={enrolment}
-                  students={family.children
-                    .concat(family.parent)
-                    .concat(family.guests)}
-                />
-              ))}
-          </Box>
-        </Box>
+        <Typography variant="h3" className={classes.heading}>
+          Previous Enrolments
+        </Typography>
+        {family.enrolments
+          .filter((enrolment) => !enrolment.session.active)
+          .map((enrolment) => (
+            <PreviousEnrolmentCard
+              key={enrolment.id}
+              enrolment={enrolment}
+              students={family.children
+                .concat(family.parent)
+                .concat(family.guests)}
+            />
+          ))}
       </Box>
 
       {isEditingFamily && (
