--- src/components/families/family-sidebar/family-sidebar-enrolment-card/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./FamilySidebarEnrolmentCard";
