+++ src/components/families/family-sidebar/previous-enrolment-card/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PreviousEnrolmentCard";
