+++ src/types/index.ts
@@ -68,11 +68,11 @@ export type Family = {
 };
 
 export type Session = {
+  active: boolean;
   fields: number[]; // array of field IDs
   id: number;
   name: string;
   start_date: string;
-  active: boolean;
 };
 
 export type Attendance = {
